pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vmec-benchmark"
version = "0.1.0"
description = "A comprehensive benchmark suite for comparing VMEC implementations"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "VMEC Benchmark Contributors"},
]
keywords = ["vmec", "benchmark", "plasma", "fusion", "stellarator"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "numpy>=1.20",
    "matplotlib>=3.3",
    "pandas>=1.3",
    "netCDF4>=1.5",
    "gitpython>=3.1",
    "pyyaml>=5.4",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.910",
    "pre-commit>=2.15",
]

[project.scripts]
vmec-benchmark = "vmec_benchmark.cli:main"

[project.urls]
"Homepage" = "https://github.com/yourusername/vmec-benchmark"
"Bug Tracker" = "https://github.com/yourusername/vmec-benchmark/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true